% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Episodes_soins.R
\encoding{UTF-8}
\name{combine_epi_soins}
\alias{combine_epi_soins}
\title{Reconstitution des épisodes de soins}
\usage{
combine_epi_soins(dt, id, debut, fin, par_cols = NULL, njours = 1)
}
\arguments{
\item{dt}{Data d'analyse.}

\item{id}{Nom de la colonne indiquant l'identifiant de l'individu.}

\item{debut}{Nom de la colonne indiquant le début de la période.}

\item{fin}{Nom de la colonne indiquant la fin de la période.}

\item{par_cols}{Nom des autres colonnes qui doivent être incluses dans l'analyse. Par exemple des codes de médicaments ou des conditions médicales.
Afin de pouvoir utiliser le paramètre \code{par_cols}, la base de données doit être sous forme long en fonction des ids et des variables incluses dans \code{par_cols}.Par défaut \code{NULL}.}

\item{njours}{Nombre de jours max entre le début et la fin précédente pour effectuer une combinaison.}
}
\description{
La fonction \code{combine_epi_soins} permet de combiner et fusionner les épisodes de soins dont les périodes se chevauchent dans le temps.
}
\examples{
 \dontrun{
 # DT <- data.table(
}
}
