% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiner_periodes.R
\encoding{UTF-8}
\name{combiner_periodes}
\alias{combiner_periodes}
\title{Combinaisons périodes d'études}
\usage{
combiner_periodes(dt, id, debut, fin, par_cols = NULL, njours = 1)
}
\arguments{
\item{dt}{Table de données.}

\item{id}{Nom de la colonne indiquant l'identifiant de l'individu.}

\item{debut}{Nom de la colonne indiquant le début de la période.}

\item{fin}{Nom de la colonne indiquant la fin de la période.}

\item{par_cols}{Nom des autres colonnes qui doivent être incluses dans l'analyse. Par exemple des codes de médicaments. Par défaut \code{NULL}.}

\item{njours}{Nombre de jours max entre le début et la fin précédente pour effectuer une combinaison.}
}
\description{
Combine, fusionne des périodes qui se chevauchent dans le temps.
}
\examples{
\dontrun{

DT <- data.table(
  ID = c(rep(1, 3), rep(2, 2)),
  DIN = c(rep(555, 3), rep(999, 2)),
  DENOM = c(rep(159, 3), rep(357, 2)),
  DatAdm = as.Date(c("2022-03-01","2022-03-15","2022-03-19",
                     "2022-06-01", "2022-06-15")),
  DatDep = as.Date(c("2022-03-25","2022-03-17","2022-03-30",
                     "2022-06-15", "2022-06-19"))
)
DT_final<-combiner_periodes(
dt=DT,
id = "ID",
debut = "DatAdm",
fin = "DatDep",
njours = 1,
par_cols = NULL #c("DIN", "DENOM")
)
}
}
