% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_index.R
\encoding{UTF-8}
\name{comorbidity.confirm_sourc_names}
\alias{comorbidity.confirm_sourc_names}
\title{SQL_comorbidity}
\usage{
comorbidity.confirm_sourc_names(confirm_sourc, val)
}
\description{
Indiquer la source de confirmation: MEDECHO, BDCU, SMOD
}
\keyword{internal}
