% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_I_SMOD_SERV_MD_CM.diagn}
\alias{query_I_SMOD_SERV_MD_CM.diagn}
\title{query_V_EPISO_SOIN_DURG_CM.diagn}
\usage{
query_I_SMOD_SERV_MD_CM.diagn(diagn)
}
\description{
Section \code{where} du code SQL où on demande des DX spécifiques et tous DX confondus.
}
\keyword{internal}
