% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_I_SMOD_SERV_MD_CM.where_I_BENF_ADR_CM}
\alias{query_I_SMOD_SERV_MD_CM.where_I_BENF_ADR_CM}
\title{query_I_SMOD_SERV_MD_CM.where_I_BENF_ADR_CM}
\usage{
query_I_SMOD_SERV_MD_CM.where_I_BENF_ADR_CM(
  benef_adr,
  debut_periode,
  fin_periode,
  date_adr
)
}
\description{
Section \code{where} du code SQL où on demande les adresses de Benef.
}
\keyword{internal}
