% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_I_SMOD_SERV_MD_CM.where_SMOD_COD_SPEC}
\alias{query_I_SMOD_SERV_MD_CM.where_SMOD_COD_SPEC}
\title{query_I_SMOD_SERV_MD_CM.where_SMOD_COD_SPEC}
\usage{
query_I_SMOD_SERV_MD_CM.where_SMOD_COD_SPEC(omni_spec)
}
\description{
Section \code{where} du code SQL où on demande les actes facturés par les omipraticiens et/ou spécialistes.
}
\keyword{internal}
