% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_I_SMOD_SERV_MD_CM_AG}
\alias{query_I_SMOD_SERV_MD_CM_AG}
\title{Code SQL: query_I_SMOD_SERV_MD_CM_AG \code{(SMOD)}}
\usage{
query_I_SMOD_SERV_MD_CM_AG(
  query,
  debut,
  fin,
  diagn,
  omni_spec,
  CodeActe,
  catg_etab
)
}
\arguments{
\item{query}{indiquant si on veut extraire des diagnostics "extraction_Dx", ou calculer le nombre d'acte "calcul_Nb_acte", ou extraire des codes d'actes "extraction_acte".}

\item{debut}{Date de début de la période d'étude au format \code{"AAAA-MM-JJ"}.}

\item{fin}{Date de fin de la période d'étude au format \code{"AAAA-MM-JJ"}.}

\item{diagn}{\code{vecteur} indiquant les codes de diagnostics \emph{CIM9} et/ou \emph{CIM10} à extraire.}

\item{omni_spec}{peut prendre la valeur \code{omni}(extraction des actes facturés par les omni: SMOD_COD_SPEC IS NULL),
\code{spec} (extraction des actes facturés par les spec: SMOD_COD_SPEC IS NOT NULL) ou \code{all} (pas de spécification).}

\item{CodeActe}{\code{vecteur} indiquant les codes d'acte d'interêt.}
}
\value{
chaîne de caractères, code SQL.
}
\description{
\code{query_I_SMOD_SERV_MD_CM_AG} est une fonction multi-tâches. Elle utilise la vue \code{I_SMOD_SERV_MD_CM} pour générer un code SQL qui permet :
\enumerate{
\item d'extraire les diagnostics pour créer par exemple une cohorte.
\item de calculer le nombre d'acte par bénéficiaire entre la date de début et de fin d'une étude.
\item d'extraire les codes d'actes, avec les caractéristiques de bénéficiaire et de l'établissement de soins.
}
}
\examples{
 \dontrun{
 **Extraction des diagnostics:**
DT<-as.data.table(odbc::dbGetQuery(conn=SQL_connexion(),
statement=query_I_SMOD_SERV_MD_CM_AG (
query="extraction_Dx",
debut="2021-04-01",
fin="2022-03-31",
diagn=c('5995\%','6180\%','6181\%','6183\%','6184\%','N810\%','N811\%','N812\%')
CodeActe,
omni_spec)
))

 **Calcul du nombre d'acte par bénéficiaire:**
DT<-as.data.table(odbc::dbGetQuery(conn=SQL_connexion(),
statement=query_I_SMOD_SERV_MD_CM_AG (
query="calcul_Nb_acte",
debut="2021-04-01",
fin="2022-03-31",
diagn,
CodeActe=c('07122', '07237', '07800', '07089', '0780'), #Si CodeActe est non spécifié, le calcul se fait sur tout les actes facturés à l'intérieur de la période d'étude.
omni_spec="all")
))

**Extraction des codes d'actes:**
DT<-as.data.table(odbc::dbGetQuery(conn=SQL_connexion(),
statement=query_I_SMOD_SERV_MD_CM_AG (
query="extraction_acte",
debut="2021-04-01",
fin="2022-03-31",
diagn,
CodeActe=c('07122', '07237', '07800', '07089', '0780'),
omni_spec="all")
))
}

}
\keyword{internal}
