% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_SQL_CodeActe.R
\encoding{UTF-8}
\name{query_SQL_CodeActe}
\alias{query_SQL_CodeActe}
\title{query_SQL_CodeActe}
\usage{
query_SQL_CodeActe(DateDebut, DateFin, CodeActe)
}
\arguments{
\item{DateDebut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{DateFin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{CodeActe}{Vecteur comprenant les codes d'acte d'interêt.}
}
\value{
Chaîne de caractères, code SQL.
}
\description{
La fonction \code{query_SQL_CodeActe} est une fonction intermédiaire qui est utilisée par la fonction \code{\link{SQL_CodeActe}} afin
d'extraire les actes entre une date de début et de fin d'étude. Pour plus de détails, veuillez-vous référer à la vignette \code{query_SQL_CodeActe.html}.
}
\examples{
 \dontrun{
query_SQL_CodeActe (DateDeb="2019-03-01",
                    DateFin="2022-05-31",
                    CodeActe=c(07122, 07237, 07800, 07089, 0780))
}

}
\keyword{internal}
