% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_SQL_CodeActe.where_ETAB_COD_CATG_ETAB_EI}
\alias{query_SQL_CodeActe.where_ETAB_COD_CATG_ETAB_EI}
\title{query_SQL_CodeActe.where_ETAB_COD_CATG_ETAB_EI}
\usage{
query_SQL_CodeActe.where_ETAB_COD_CATG_ETAB_EI(catg_etab)
}
\description{
Section \code{where} du code SQL où on demande les visites en ambulatoire.
}
\keyword{internal}
