% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_SQL_CodeActe.where_SMOD_COD_ACTE}
\alias{query_SQL_CodeActe.where_SMOD_COD_ACTE}
\title{query_SQL_CodeActe.where_SMOD_COD_ACTE}
\usage{
query_SQL_CodeActe.where_SMOD_COD_ACTE(CodeActe)
}
\description{
Section \code{where} du code SQL où on demande les actes.
}
\keyword{internal}
