% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_SQL_CodeActe.where_SMOD_COD_STA_DECIS}
\alias{query_SQL_CodeActe.where_SMOD_COD_STA_DECIS}
\title{query_SQL_CodeActe.where_SMOD_COD_STA_DECIS}
\usage{
query_SQL_CodeActe.where_SMOD_COD_STA_DECIS(code_stat_decis)
}
\description{
Section \code{where} du code SQL où on demande le Code de statu de décision.
}
\keyword{internal}
