% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_DIAGN_SEJ_HOSP_CM.R,
%   R/query_V_DIAGN_SEJ_HOSP_CM_AG.R
\encoding{UTF-8}
\name{query_V_DIAGN_SEJ_HOSP_CM.where_typ_diagn}
\alias{query_V_DIAGN_SEJ_HOSP_CM.where_typ_diagn}
\title{query_V_DIAGN_SEJ_HOSP_CM.where_typ_diagn}
\usage{
query_V_DIAGN_SEJ_HOSP_CM.where_typ_diagn(typ_diagn)

query_V_DIAGN_SEJ_HOSP_CM.where_typ_diagn(typ_diagn)
}
\description{
Section \code{where} du code SQL où on demande les types de codes de diagnostics. Voir variable \code{SHOP_TYP_DIAGN_SEJ_HOSP}.

Section \code{where} du code SQL où on demande les types de codes de diagnostics. Voir variable \code{SHOP_TYP_DIAGN_SEJ_HOSP}.
}
\keyword{internal}
