% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_EPISO_SOIN_DURG_CM.R,
%   R/query_V_EPISO_SOIN_DURG_CM_AG.R
\encoding{UTF-8}
\name{query_V_EPISO_SOIN_DURG_CM.date_dx_var}
\alias{query_V_EPISO_SOIN_DURG_CM.date_dx_var}
\title{query_V_EPISO_SOIN_DURG_CM}
\usage{
query_V_EPISO_SOIN_DURG_CM.date_dx_var(date_dx_var)

query_V_EPISO_SOIN_DURG_CM.date_dx_var(date_dx_var)
}
\description{
Retourne le nom de la variable à utiliser selon la valeur de la variable \code{date_dx_var}.

Retourne le nom de la variable à utiliser selon la valeur de la variable \code{date_dx_var}.
}
\keyword{internal}
