% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_FICH_ID_BEN_CM.R
\encoding{UTF-8}
\name{query_V_FICH_ID_BEN_CM}
\alias{query_V_FICH_ID_BEN_CM}
\title{Code SQL: Extraction des caractéristiques du bénéficiaire}
\usage{
query_V_FICH_ID_BEN_CM(debut_periode, fin_periode, date_age)
}
\arguments{
\item{debut_periode}{Date de début de la période d'étude.}

\item{fin_periode}{Date de fin de la période d'étude.}

\item{date_age}{\code{"AAAA-MM-JJ"} indique la date à laquelle l'âge des bénéficaires est calculé.}
}
\value{
chaîne de caractères, code SQL.
}
\description{
Générateur de code SQL pour l'extraction des variables suivante: BENF_NO_INDIV_BEN_BANLS, BENF_COD_SEXE, BENF_DAT_NAISS, BENF_DAT_DECES de la vue \code{V_FICH_ID_BEN_CM}.
L'âge est calculé à une date fixe spécifiée par l'argement \code{date_age}.
}
\examples{
\dontrun{
DT1<-as.data.table(odbc::dbGetQuery(conn=SQL_connexion(),
statement=query_V_FICH_ID_BEN_CM(debut_periode="2022-02-01", fin_periode="2022-03-31", date_age="2022-03-31")
}


}
\keyword{internal}
