% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_SEJ_HOSP_CM.R
\encoding{UTF-8}
\name{query_V_SEJ_HOSP_CM}
\alias{query_V_SEJ_HOSP_CM}
\title{Code SQL: Extraction de diagnostics dans BDCU}
\usage{
query_V_SEJ_HOSP_CM(
  query,
  debut_periode,
  fin_periode,
  date_dx_var,
  benef_adr,
  date_adr
)
}
\arguments{
\item{date_dx_var}{\code{'admis'} ou \code{'depar'}. Indique si on utilise la date d'admission ou la date de départ comme date indexe pour l'étude.}

\item{debut_cohort}{Date de début de la période d'étude.}

\item{fin_cohort}{Date de fin de la période d'étude.}
}
\value{
chaîne de caractères, code SQL.
}
\description{
Générateur de code SQL pour l'extraction de diagnostics de la vue \code{V_SEJ_HOSP_CM}.
}
\keyword{internal}
