% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_SEJ_SERV_HOSP_CM_AG.R
\encoding{UTF-8}
\name{query_V_SEJ_SERV_HOSP_CM.where_diagn_AG}
\alias{query_V_SEJ_SERV_HOSP_CM.where_diagn_AG}
\title{query_V_DIAGN_SEJ_HOSP_CM.where_diagn_AG}
\usage{
query_V_SEJ_SERV_HOSP_CM.where_diagn_AG(diagn)
}
\description{
Section \code{where} du code SQL où on demande les codes de diagnostics.
}
\keyword{internal}
