% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_V_SEJ_SERV_HOSP_CM_AG.R
\encoding{UTF-8}
\name{query_V_SEJ_SERV_HOSP_CM_AG}
\alias{query_V_SEJ_SERV_HOSP_CM_AG}
\title{Code SQL: Extraction de diagnostics dans MEDECHO-sej}
\usage{
query_V_SEJ_SERV_HOSP_CM_AG(
  query,
  debut_periode,
  fin_periode,
  debut_cohort,
  fin_cohort,
  diagn,
  date_dx_var
)
}
\arguments{
\item{debut_cohort}{Date de début de la période d'étude.}

\item{fin_cohort}{Date de fin de la période d'étude.}

\item{diagn}{\code{vector} indiquant les codes de diagnostics \emph{CIM9} et/ou \emph{CIM10} à extraire.}

\item{date_dx_var}{\code{'admis'} ou \code{'depar'}. Indique si on utilise la date d'admission ou la date de départ comme date de diagnostic pour l'étude.}
}
\value{
chaîne de caractères, code SQL.
}
\description{
Générateur de code SQL pour l'extraction de diagnostics de la vue \code{V_SEJ_SERV_HOSP_CM}.
}
\keyword{internal}
