% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteGeneriqueBDCA_shiny.R
\encoding{UTF-8}
\name{RequeteGeneriqueBDCA_shiny}
\alias{RequeteGeneriqueBDCA_shiny}
\title{RequeteGeneriqueBDCA_shiny}
\usage{
RequeteGeneriqueBDCA_shiny()
}
\description{
L'application \code{RequeteGeneriqueBDCA_shiny} est destinée aux utilisateurs.

#' Elle permet: \cr
\enumerate{
\item d'extraire les variables pertinantes de différentes vues de données.\cr
\item de creéer une cohorte en utilisant les Dx cim9 et cim10.\cr
\item d'appliquer une standardisation directe ou indirect. \cr
}
}
