% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins-rstudio.R
\encoding{UTF-8}
\name{RequeteGeneriqueBDCA_shiny_addin}
\alias{RequeteGeneriqueBDCA_shiny_addin}
\title{Démarrer RequeteGeneriqueBDCA_shiny}
\usage{
RequeteGeneriqueBDCA_shiny_addin()
}
\value{
\link{RequeteGeneriqueBDCA_shiny}
}
\description{
Démarrer RequeteGeneriqueBDCA_shiny
}
\keyword{internal}
