% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_CodeActe.R
\encoding{UTF-8}
\name{SQL_CodeActe}
\alias{SQL_CodeActe}
\title{SQL_CodeActe}
\usage{
SQL_CodeActe(conn, statement, DateDebut, DateFin, CodeActe)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}.}

\item{statement}{Fait appel à la fonction \code{query_SQL_CodeActe}. Voir \code{\link{query_SQL_CodeActe}}.}

\item{DateDebut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{DateFin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{CodeActe}{Vecteur comprenant les codes d'acte d'interêt.}
}
\value{
\code{data.table} de 31 variables :
\itemize{
\item \strong{\code{BenBanls}} : Numéro d'identification du bénéficiaire.
\item \strong{\code{AnCivil}} : L'année civil de l'acte.
\item \strong{\code{DateActe}} : La date de l'acte au format \code{AAAA-MM-JJ}.
\item \strong{\code{CodeActe}} : Les actes demandés dans l'argument \code{CodeActe}.
\item \strong{\code{CoutActe}} : Le coût de de l'acte.
\item \strong{\code{DxActe}}: Num de diagnostique associé à l'acte.
\item \strong{\code{LieuDisp}}: Code du lieu de dispensation d'un service (C:Service rendu en cabinet ou à domicile; E:Service rendu en établissement; X:Indéterminé).
\item \strong{\code{SecActiv}}: Identifie de façon unique chacun des secteurs d'activités dans lesquels les différents établissements peuvent oeuvrer.
\item \strong{\code{Num_ETAB_USUEL}}: Numéro de l'établissement.
\item \strong{\code{CodEntente}}: Code d'entente. Il utile pour associer les dispensateurs à un groupe spécifique.
\item \strong{\code{Sexe}}: Sexe du bénéficiaire.
\item \strong{\code{DatNais}}: Date de naissance du bénéficiaire au format \code{AAAA-MM-JJ}.
\item \strong{\code{datdeces}}: Date de décès du bénéficiaire au format \code{AAAA-MM-JJ}.
\item \strong{\code{Age}}: Age du bénéficiaire calculé jusqu'à la date de l'acte.
\item \strong{\code{NumDispSp}}: Numéro de dispensateur.
\item \strong{\code{SPDisps}}: Spécialité de dispensateur.
\item \strong{\code{NumDispRef}}: Numéro de dispensateur référent.
\item \strong{\code{SPDispRefs}}: Spécialité de dispensateur référent.
\item \strong{\code{NO_ETAB}}: Numéro de l'établissement.
\item \strong{\code{Grp_Cat_Etab}}: Catégorie de l'établissement.
\item \strong{\code{Cat_Etab}} Catégorie de l'établissement: .
\item \strong{\code{Nom_Grp_Cat_Etab}}: Nom de la catégorie de l'établissement.
\item \strong{\code{NOM_ETAB}}: Nom de l'établissement.
\item \strong{\code{RSS_Benf}}: Régions sociosanitaires (RSS) du bénéficiaire.
\item \strong{\code{RLS_Benef}}: Réseaux locaux de services (RLS) du bénéficiaire.
\item \strong{\code{NomRLS_Benef}}: Nom réseaux locaux de services (RLS) du bénéficiaire.
\item \strong{\code{NomRSS_Benef}}: Nom régions sociosanitaires (RSS) du bénéficiaire.
\item \strong{\code{RSS_Etab}}: Régions sociosanitaires (RSS) de l'établissement.
\item \strong{\code{RLS_Etab}}: Réseaux locaux de services (RLS) de l'établissement.
\item \strong{\code{NomRLS_Etab}}: Nom réseaux locaux de services (RLS) de l'établissement.
\item \strong{\code{NomRSS_Etab}}: Nom régions sociosanitaires (RSS) de l'établissement.
}
}
\description{
La fonction \code{SQL_CodeActe} permet d'extraire les actes entre une date de début et de fin d'étude.
La fonction \code{SQL_CodeActe} utilise la fonction générique \code{query_SQL_CodeActe}. Voir \code{\link{query_SQL_CodeActe}}
}
\examples{
\dontrun{
tmp<-SQL_CodeActe(conn=conn,
                 statement=query_SQL_CodeActe(),
                 DateDebut="2019-03-01",
                 DateFin="2019-05-31",
                 CodeActe=c(07122, 07237, 07800, 07089, 0780))
}

}
