% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_index.R
\encoding{UTF-8}
\name{SQL_comorbidity.exclu_diab_gross}
\alias{SQL_comorbidity.exclu_diab_gross}
\title{SQL_comorbidity}
\usage{
SQL_comorbidity.exclu_diab_gross(
  conn,
  dt,
  CIM,
  dt_source,
  dt_desc,
  date_dx_var,
  typ_diagn,
  verbose
)
}
\description{
Exclusion des cas de diabète et d'hypertension gestationnel.
}
\keyword{internal}
