% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_index.R
\encoding{UTF-8}
\name{SQL_obstetric}
\alias{SQL_obstetric}
\title{Extraction - Codes diagn obstetriques}
\usage{
SQL_obstetric(
  conn = SQL_connexion(),
  cohort,
  debut,
  fin,
  CIM = c("CIM9", "CIM10"),
  dt_source = c("V_DIAGN_SEJ_HOSP_CM", "V_SEJ_SERV_HOSP_CM", "V_EPISO_SOIN_DURG_CM",
    "I_SMOD_SERV_MD_CM"),
  dt_desc = list(V_DIAGN_SEJ_HOSP_CM = "MED-ECHO", V_SEJ_SERV_HOSP_CM = "MED-ECHO",
    V_EPISO_SOIN_DURG_CM = "BDCU", I_SMOD_SERV_MD_CM = "SMOD"),
  date_dx_var = "depar",
  typ_diagn = c("A", "P", "S"),
  verbose = TRUE
)
}
\description{
Extraction des événements obstétriques.
}
