% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_I_SMOD_SERV_MD_CM_shiny.R
\encoding{UTF-8}
\name{user_I_SMOD_SERV_MD_CM_shiny}
\alias{user_I_SMOD_SERV_MD_CM_shiny}
\title{user_I_SMOD_SERV_MD_CM_shiny}
\usage{
user_I_SMOD_SERV_MD_CM_shiny()
}
\description{
La fonction \code{user_I_SMOD_SERV_MD_CM_shiny} est destinée aux utilisateurs. Elle permet d'exécuter la rêquette SQL retournée par la fonction \code{query_I_SMOD_SERV_MD_CM_AG}.\cr

#' Elle permet: \cr
\enumerate{
\item d'extraire les variables pertinantes de la base de données SMOD, jumlées avec d'autres bases de données pour rajouter les caractéristiques de dispensateur, du bénéficiaire et de l'établissement de soins.\cr
\item d'extraire les variables pertinantes de la base de données SMOD et d'effectuer des analyses discriptives afin d'orienter la prise de décision.\cr
}

La fonction \code{user_I_SMOD_SERV_MD_CM_shiny} utilise plusieurs autres fonctions génériques telles que:\cr

\code{SQL_reperage_cond_med} pour la création de la cohorte. Voir \code{\link{SQL_reperage_cond_med}}\cr
\code{query_V_FICH_ID_BEN_CM} pour l'extraction des caractéristiques de bénéficiaire. Voir \code{\link{query_V_FICH_ID_BEN_CM}}
}
