% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins-rstudio.R
\encoding{UTF-8}
\name{user_I_SMOD_SERV_MD_CM_shiny_addin}
\alias{user_I_SMOD_SERV_MD_CM_shiny_addin}
\title{Démarrer user_I_SMOD_SERV_MD_CM_shiny}
\usage{
user_I_SMOD_SERV_MD_CM_shiny_addin()
}
\value{
\link{user_I_SMOD_SERV_MD_CM_shiny}
}
\description{
Démarrer user_I_SMOD_SERV_MD_CM_shiny
}
\keyword{internal}
