% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qu.R
\encoding{UTF-8}
\name{qu}
\alias{qu}
\title{qu}
\usage{
qu(Dx)
}
\arguments{
\item{Dx}{Un vecteur incluant les codes d'acte ou les diagnostics.}
}
\value{
Chaîne de caractères sans les quotes ("").
}
\description{
La fonction \code{qu} permet de mettre de formater les codes d'acte et les diagnostics pour qu'ils soient laisible dans les rêquetes SQL.
}
\examples{
 \dontrun{
Dx<-c(07122, 07237, 07800, 07089, 0780)
qu(Dx)
}
}
\keyword{internal}
