% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_I_SMOD_SERV_MD_CM_AG.R
\encoding{UTF-8}
\name{query_I_SMOD_SERV_MD_CM_AG}
\alias{query_I_SMOD_SERV_MD_CM_AG}
\title{Code SQL: query_I_SMOD_SERV_MD_CM_AG \code{(SMOD)}}
\usage{
query_I_SMOD_SERV_MD_CM_AG(
  query,
  debut,
  fin,
  diagn,
  omni_spec,
  CodeActe,
  catg_etab,
  code_stat_decis
)
}
\arguments{
\item{query}{indiquant si on veut effectuer 1) une extraction brute des vaiables pertinantes dans SMOD "extraction_SMOD", 2) une extraction brute des vaiables pertinantes dans SMOD jumlées avec d'autres bases de données "extraction_SMOD_combin",
3) une extraction des diagnostics "extraction_Dx", 4) une extraction des dates de visites, ou 5) calculer le nombre d'acte "calcul_Nb_acte".}

\item{debut}{Date de début de la période d'étude au format \code{"AAAA-MM-JJ"}.}

\item{fin}{Date de fin de la période d'étude au format \code{"AAAA-MM-JJ"}.}

\item{diagn}{\code{vecteur} indiquant les codes de diagnostics \emph{CIM9} et/ou \emph{CIM10} à extraire.}

\item{omni_spec}{peut prendre la valeur \code{omni}(extraction des actes facturés par les omni: SMOD_COD_SPEC IS NULL),
\code{spec} (extraction des actes facturés par les spec: SMOD_COD_SPEC IS NOT NULL) ou \code{all} (pas de spécification).}

\item{CodeActe}{\code{vecteur} indiquant les codes d'acte d'interêt.}

\item{catg_etab}{peut prendre la valeur \code{ambulatoire} (pour les actes facturés en ambulatoire) ou \code{all} (pas de spécification).}

\item{code_stat_decis}{peut prendre la valeur \code{PAY-PPY}, \code{PAY}, ou \code{PPY}}
}
\value{
chaîne de caractères, code SQL.
}
\description{
\code{query_I_SMOD_SERV_MD_CM_AG} est une fonction multi-tâches. Elle utilise principalement la vue \code{I_SMOD_SERV_MD_CM} jumlée avec d'autres vues pour générer un code SQL qui permet :\cr
\enumerate{
\item d'extraire les variables pertinantes de la base de données SMOD.\cr
\item d'extraire les variables pertinantes de la base de données SMOD, jumlées avec d'autres bases de données pour rajouter les caractéristiques de dispensateur, du bénéficiaire et de l'établissement de soins.\cr
\item d'extraire les diagnostics pour créer par exemple une cohorte.\cr
\item d'extraire les dates de visites (services).
\item de calculer le nombre d'acte par bénéficiaire entre la date de début et de fin d'une étude.
}
}
\keyword{internal}
