% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_SQL_CodeActe.R
\encoding{UTF-8}
\name{query_SQL_CodeActe}
\alias{query_SQL_CodeActe}
\title{Requête SQL: query_SQL_CodeActe}
\usage{
query_SQL_CodeActe(debut, fin, CodeActe, omni_spec)
}
\arguments{
\item{debut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{CodeActe}{Vecteur comprenant les codes d'acte d'interêt.}

\item{omni_spec}{peut prendre la valeur \code{omni},\code{spec} ou \code{all}. Si omi_spec == "omni" (AND SMOD_COD_SPEC IS NULL),Si omi_spec == "spec" (AND SMOD_COD_SPEC IS NOT NULL), et si omni_spec=="all" (\verb{/*AND SMOD_COD_SPEC IS NOT NULL*/})}
}
\value{
Chaîne de caractères, code SQL.
}
\description{
La fonction \code{query_SQL_CodeActe} est une fonction intermédiaire qui est utilisée par la fonction \code{\link{SQL_CodeActe}} afin
d'extraire les actes entre une date de début et de fin d'étude avec des données sur le bénéficiaire ainsi que l'établissement de soins. Pour plus de détails, veuillez-vous référer à la vignette \code{query_SQL_CodeActe.html}.
}
\examples{
 \dontrun{
DT<-as.data.table(odbc::dbGetQuery(conn=SQL_connexion(),
statement=query_SQL_CodeActe (debut="2019-04-01",fin="2022-03-31",
CodeActe=c(07122, 07237, 07800, 07089, 0780),
omni_spec="all")
))
}

}
\keyword{internal}
