% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL_CodeActe.R
\encoding{UTF-8}
\name{SQL_CodeActe}
\alias{SQL_CodeActe}
\title{SQL_CodeActe}
\usage{
SQL_CodeActe(
  conn = SQL_connexion(),
  debut,
  fin,
  CodeActe,
  Dx_table,
  CIM,
  by_Dx,
  nDx,
  date_dx_var,
  n1,
  n2,
  date_age,
  verbose
)
}
\arguments{
\item{conn}{Variable contenant la connexion entre R et Teradata. Voir \code{\link{SQL_connexion}}.}

\item{debut}{Date de début de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{fin}{Date de fin de la période d'étude au format \code{AAAA-MM-JJ}.}

\item{CodeActe}{Vecteur comprenant les codes d'acte d'interêt.}

\item{Dx_table}{\code{list} contenant les codes à analyser ou le nom d'une table qui est inclue dans la librairie.}

\item{CIM}{\code{'CIM9'}, \code{'CIM10'} ou les deux. Permet de filtrer les codes de diagnostics selon le numéro de révision de la \emph{Classification statistique internationale des maladies et des problèmes de santé connexes} (CIM).}

\item{by_Dx}{\code{TRUE} ou \code{FALSE}. Distinction entre les diagnostics (\code{TRUE}) ou pas (\code{FALSE}). Si \code{TRUE}, on considère chaque élément de \code{Dx_table} où chaque élément peut contenir plusieurs codes. Le nombre d'éléments sera donc le nombre maximal d'observations (lignes résultats) par individu.\cr Si \code{FALSE}, on considère tous les éléments de \code{Dx_Table} comme un seul, on aura donc au maximum une ligne résultat par individu.\cr Voir Détails.}

\item{nDx}{Nombre de diagnostics qui doivent confirmer la 1re date. Par défaut 1.}

\item{date_dx_var}{\verb{'admis} ou \code{'depar'}. Indique si on utilise la date d'admission ou la date de départ comme date de diagnostic pour l'étude dans les vues V_DIAGN_SEJ_HOSP_CM, V_SEJ_SERV_HOSP_CM et V_EPISO_SOIN_DURG_CM.}

\item{n1, n2}{Nombre de jours requis pour confirmer le diagnostic}

\item{date_age}{\code{"AAAA-MM-JJ"} indique la date à laquelle l'âge des bénéficaires est calculé.}
}
\value{
\code{data.table} de 31 variables :
\itemize{
\item \strong{\code{BenBanls}} : Numéro d'identification du bénéficiaire.
\item \strong{\code{AnCivil}} : L'année civil de l'acte.
\item \strong{\code{DateActe}} : La date de l'acte au format \code{AAAA-MM-JJ}.
\item \strong{\code{CodeActe}} : Les actes demandés dans l'argument \code{CodeActe}.
\item \strong{\code{CoutActe}} : Le coût de de l'acte.
\item \strong{\code{DxActe}}: Num de diagnostique associé à l'acte.
\item \strong{\code{LieuDisp}}: Code du lieu de dispensation d'un service (C:Service rendu en cabinet ou à domicile; E:Service rendu en établissement; X:Indéterminé).
\item \strong{\code{SecActiv}}: Identifie de façon unique chacun des secteurs d'activités dans lesquels les différents établissements peuvent oeuvrer.
\item \strong{\code{Num_ETAB_USUEL}}: Numéro de l'établissement.
\item \strong{\code{CodEntente}}: Code d'entente. Il utile pour associer les dispensateurs à un groupe spécifique.
\item \strong{\code{Sexe}}: Sexe du bénéficiaire.
\item \strong{\code{DatNais}}: Date de naissance du bénéficiaire au format \code{AAAA-MM-JJ}.
\item \strong{\code{datdeces}}: Date de décès du bénéficiaire au format \code{AAAA-MM-JJ}.
\item \strong{\code{Age}}: Age du bénéficiaire calculé jusqu'à la date de l'acte.
\item \strong{\code{NumDispSp}}: Numéro de dispensateur.
\item \strong{\code{SPDisps}}: Spécialité de dispensateur.
\item \strong{\code{NumDispRef}}: Numéro de dispensateur référent.
\item \strong{\code{SPDispRefs}}: Spécialité de dispensateur référent.
\item \strong{\code{NO_ETAB}}: Numéro de l'établissement.
\item \strong{\code{Grp_Cat_Etab}}: Catégorie de l'établissement.
\item \strong{\code{Cat_Etab}} Catégorie de l'établissement: .
\item \strong{\code{Nom_Grp_Cat_Etab}}: Nom de la catégorie de l'établissement.
\item \strong{\code{NOM_ETAB}}: Nom de l'établissement.
\item \strong{\code{RSS_Benf}}: Régions sociosanitaires (RSS) du bénéficiaire.
\item \strong{\code{RLS_Benef}}: Réseaux locaux de services (RLS) du bénéficiaire.
\item \strong{\code{NomRLS_Benef}}: Nom réseaux locaux de services (RLS) du bénéficiaire.
\item \strong{\code{NomRSS_Benef}}: Nom régions sociosanitaires (RSS) du bénéficiaire.
\item \strong{\code{RSS_Etab}}: Régions sociosanitaires (RSS) de l'établissement.
\item \strong{\code{RLS_Etab}}: Réseaux locaux de services (RLS) de l'établissement.
\item \strong{\code{NomRLS_Etab}}: Nom réseaux locaux de services (RLS) de l'établissement.
\item \strong{\code{NomRSS_Etab}}: Nom régions sociosanitaires (RSS) de l'établissement.
}
}
\description{
La fonction \code{SQL_CodeActe} permet d'extraire les actes entre une date de début et de fin d'étude.
La fonction \code{SQL_CodeActe} utilise plusieurs autres fonctions génériques telles que:
\code{SQL_reperage_cond_med} Voir \code{\link{SQL_reperage_cond_med}}, \code{query_SQL_CodeActe} Voir \code{\link{query_SQL_CodeActe}}, et \code{query_V_FICH_ID_BEN_CM} Voir \code{\link{query_V_FICH_ID_BEN_CM}}
}
\examples{
\dontrun{
DT_final<-SQL_CodeActe(conn=SQL_connexion(),
                      debut = "2021-04-01",
                      fin = "2022-03-31",
                      CodeActe = c('06452','06148','06154','06265','06251','06216','06208','06164','06270','06276','06274','06273','06913','06099',
                                   '06400','06405','06406','06408','06420','06410','06411','06414','6421','6426','20017'),
                      Dx_table = list(Prolapsus = list(CIM9 = c('5995','6180','6181','6183','6184','6185','6186','6187','6188','6189','6204'),
                                                       CIM10 = c('N810','N811','N812','N813','N814','N815','N816','N818','N819','N993','N834','O345'))),

                      CIM =  c("CIM9", "CIM10"),
                      by_Dx = FALSE,
                      date_dx_var = "admis",
                      n1 = 30, n2 = 730,
                      nDx=0 ,
                      date_age="2022-03-31",
                      verbose = TRUE)
}
}
